Move vmwithcontracts out of examples
