+++ x/contracts/vm/.goreleaser.yml
