+++ x/contracts/vm/actions/transfer_test.go
@@ -13,9 +13,9 @@ import (
 	"github.com/ava-labs/hypersdk/chain/chaintest"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/codec/codectest"
-	"github.com/ava-labs/hypersdk/examples/vmwithcontracts/storage"
 	"github.com/ava-labs/hypersdk/state"
 	"github.com/ava-labs/hypersdk/state/tstate"
+	"github.com/ava-labs/hypersdk/x/contracts/vm/storage"
 )
 
 func TestTransferAction(t *testing.T) {
