+++ x/contracts/vm/cmd/vmwithcontracts-cli/cmd/errors.go
