+++ x/contracts/vm/cmd/vmwithcontracts-cli/cmd/genesis.go
