+++ x/contracts/vm/cmd/vmwithcontracts-cli/cmd/key.go
