+++ x/contracts/vm/cmd/vmwithcontracts-cli/cmd/root.go
