+++ x/contracts/vm/scripts/build.release.sh
