+++ x/contracts/vm/scripts/build.sh
