+++ x/contracts/vm/scripts/fix.lint.sh
