+++ x/contracts/vm/scripts/lint.sh
