+++ x/contracts/vm/scripts/run.sh
