+++ x/contracts/vm/scripts/stop.sh
