+++ x/contracts/vm/storage/state_manager.go
